requirements.txt
streamlit
google-genai
mcp
fastmcp
pandas
plotly
python-dotenv
psycopg2-binary